# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoLearn API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: activities
    description: Sustainability activity operations
  - name: users
    description: User profile operations
  - name: leaderboard
    description: Leaderboard and ranking operations
  - name: admin
    description: Admin panel operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout the current user
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities for the current user
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: submitActivity
      tags: [activities]
      summary: Submit a new activity with image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SubmitActivityBody"
      responses:
        "201":
          description: Activity submitted and verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get an activity by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /activities/stats:
    get:
      operationId: getActivityStats
      tags: [activities]
      summary: Get activity statistics for the current user
      responses:
        "200":
          description: Activity statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get the current user's full profile with badges and stats
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/badges:
    get:
      operationId: getUserBadges
      tags: [users]
      summary: Get badges for the current user
      responses:
        "200":
          description: List of user badges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Badge"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get the top users leaderboard
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  /admin/activities:
    get:
      operationId: adminListActivities
      tags: [admin]
      summary: Admin - list all user submissions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: All activity submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin - get overall platform statistics
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/activities/{id}/override:
    patch:
      operationId: adminOverrideActivity
      tags: [admin]
      summary: Admin - override activity verification result
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOverrideBody"
      responses:
        "200":
          description: Activity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, faculty, admin]
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        points:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - points
        - streak
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    Activity:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [valid, invalid, pending]
        confidence:
          type: ["number", "null"]
        predictedLabel:
          type: ["string", "null"]
        pointsAwarded:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - category
        - status
        - pointsAwarded
        - createdAt

    ActivityListResponse:
      type: object
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - activities
        - total
        - page
        - limit

    SubmitActivityBody:
      type: object
      properties:
        category:
          type: string
          enum: [tree_planting, waste_cleaning, recycling, composting, energy_saving]
        description:
          type: string
        image:
          type: string
          format: binary
      required:
        - category
        - image

    ActivityStats:
      type: object
      properties:
        totalActivities:
          type: integer
        validActivities:
          type: integer
        invalidActivities:
          type: integer
        totalPoints:
          type: integer
        currentStreak:
          type: integer
        categoryCounts:
          type: object
          additionalProperties:
            type: integer
        recentActivities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
      required:
        - totalActivities
        - validActivities
        - invalidActivities
        - totalPoints
        - currentStreak
        - categoryCounts
        - recentActivities

    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        earnedAt:
          type: ["string", "null"]
        earned:
          type: boolean
      required:
        - id
        - name
        - description
        - icon
        - earned

    UserProfile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        stats:
          $ref: "#/components/schemas/ActivityStats"
        rank:
          type: ["integer", "null"]
      required:
        - user
        - badges
        - stats

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        points:
          type: integer
        streak:
          type: integer
        validActivities:
          type: integer
        badges:
          type: integer
      required:
        - rank
        - userId
        - name
        - points
        - streak
        - validActivities
        - badges

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalActivities:
          type: integer
        validActivities:
          type: integer
        invalidActivities:
          type: integer
        pendingActivities:
          type: integer
        totalPointsAwarded:
          type: integer
        categoryBreakdown:
          type: object
          additionalProperties:
            type: integer
        recentActivities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
      required:
        - totalUsers
        - totalActivities
        - validActivities
        - invalidActivities
        - pendingActivities
        - totalPointsAwarded
        - categoryBreakdown
        - recentActivities

    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit

    AdminOverrideBody:
      type: object
      properties:
        status:
          type: string
          enum: [valid, invalid]
        notes:
          type: string
      required:
        - status